requirements.txt
APScheduler==3.6.3
atomicwrites==1.4.0
attrs==20.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
cryptography==3.1.1
decorator==4.4.2
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.1.1
python-dotenv==0.14.0
python-telegram-bot==13.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
tornado==6.0.4
tzlocal==2.1
urllib3==1.25.11
zipp==3.3.2
